requirements.txt
Flask
python-dotenv
requests
scikit-learn
pandas
psycopg2-binary
Flask-SQLAlchemy
psutil